separate out producer logic and handle none values
